requirements.txt
streamlit
streamlit-lottie
python-dotenv
google-generativeai
youtube-search-python
requests
streamlit>=1.33,<2
google-generativeai>=0.7.0
python-dotenv>=1.0.1

